ast: Fix compiler warnings reported by Russell Owen

+++ libraries/ast/unit.c
@@ -25,12 +25,12 @@
 *     License as published by the Free Software Foundation, either
 *     version 3 of the License, or (at your option) any later
 *     version.
-*     
+*
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Lesser General Public License for more details.
-*     
+*
 *     You should have received a copy of the GNU Lesser General
 *     License along with this program.  If not, see
 *     <http://www.gnu.org/licenses/>.
@@ -913,7 +913,7 @@ static int ComplicateTree( UnitNode **node, int *status ) {
       if( con != AST__BAD && (*node)->arg[ 1 ]->opcode == OP_LN ) {
          fk = 10.0*con*log( 10.0 );
          k = NINT(fk);
-         if( EQUAL(fk,k) ) {
+         if( EQUAL(fk,((double)k)) ) {
             newnode = NewNode( NULL, OP_LOG, status );
             con = k/10.0;
          } else {
